Refactor main.js to separate out Cube and Dimensions classes
The classes Cube and Dimensions were removed from the main.js file and each was placed in its own corresponding file, cube.js and dimensions.js respectively. This change was made to improve readability and maintainability by keeping the different JavaScript classes in separate files. This way, it'll be easier to develop and debug each class individually.
