+++ src/dimensions.js
@@ -1,0 +1,7 @@
+export class Dimensions {
+    constructor(height, width, depth) {
+        this.height = height;
+        this.width = width;
+        this.depth = depth;
+    }
+}
